Hopscotch Algorithm
This includes the basis of a "hopscotch algorithm" which is helpful for
drawing-programs with controls that snap to points. The example a
drawing program that draws lines by filling in cells. The issue is
that browsers debounce user mouse input and, if the user is quick
enough, they can skip over in-between points and draw wonky lines.
This resolves that issue.