kami2-solver: Optimize for checkerboard like graphs
It only seemed to halve the number of iterations of the given graphs
(you can see the changes to the assert in test_solver.py), and that
account for the increase in the time of each iteration. I tested with a
moderately difficult example (n_nodes=41) and it at least decreased from
290 secs to 275 secs; a 5% decrease, but not as much as I was hoping
for. Granted, that graph was not very checkerboard-y. We'll see how the
optimizations work for the main puzzle I'm struggling with.